+++ tests/units/hash_tables/maps/open_address_map_st_test.cpp
@@ -228,6 +228,28 @@ TEMPLATE_TEST_CASE("hash_tables::maps::open_address_map_st", "",
         CHECK(map.at(key2) == mapped2);
     }
 
+    SECTION("get_or_create_with_factory") {
+        map_type map;
+
+        constexpr int mapped = 123;
+        const auto key = std::to_string(mapped);
+        CHECK(map.emplace(key, mapped));
+        CHECK(map.size() == 1);
+        CHECK(map.at(key) == mapped);
+
+        const int mapped2 = 12345;
+        CHECK(map.get_or_create_with_factory(
+                  key, [&mapped2] { return mapped2; }) == mapped);
+        CHECK(map.size() == 1);
+        CHECK(map.at(key) == mapped);
+
+        const auto key2 = std::to_string(mapped2);
+        CHECK(map.get_or_create_with_factory(
+                  key2, [&mapped2] { return mapped2; }) == mapped2);
+        CHECK(map.size() == 2);
+        CHECK(map.at(key2) == mapped2);
+    }
+
     SECTION("operator[] (non const)") {
         map_type map;