+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # Change Log
 
-## v0.2.0 (on development)
+## v0.3.0 (on development)
 
-**TODO**: Abstract here.
+### New Features
+
+- Support of factory function when creating values conditionally ([Issue #13](https://gitlab.com/MusicScience37/cpp-hash-tables/-/issues/13)).
+
+## v0.2.0 (2022/5/6)
+
+Concurrent hash table and map, and some improvements.
 
 ### New Features
 
 - Cache of hash numbers ([Issue #11](https://gitlab.com/MusicScience37/cpp-hash-tables/-/issues/11)).
+- Concurrent hash table and map ([Issue #8](https://gitlab.com/MusicScience37/cpp-hash-tables/-/issues/8), [Issue #10](https://gitlab.com/MusicScience37/cpp-hash-tables/-/issues/10)).
 
 ### Improvements